+++ ReserveIt/Controllers/ReservationsController.cs
@@ -18,38 +18,58 @@ namespace ReserveIt.Controllers
         [HttpGet]
         public ActionResult Index()
         {
-            var reservations = db.Reservations.Include(r => r.User);
-            return View(reservations.ToList());
+            if (Session["accessLevel"] != null)
+            {
+                var reservations = db.Reservations.Include(r => r.User);
+                return View(reservations.ToList());
+            }
+
+            return RedirectToAction("Index", "Admin");
         }
 
         [HttpPost]
         public ActionResult Index(AdminReservationSearchData data)
         {
-            int stayLength = (data.CheckOut - data.CheckIn).Days;
-            var reservations = db.Reservations.Where(x => x.CheckIn == data.CheckIn && x.StayLength == stayLength).Include(r => r.User);
-            return View(reservations.ToList());
+            if (Session["accessLevel"] != null)
+            {
+                int stayLength = (data.CheckOut - data.CheckIn).Days;
+                var reservations = db.Reservations.Where(x => x.CheckIn == data.CheckIn && x.StayLength == stayLength).Include(r => r.User);
+                return View(reservations.ToList());
+            }
+
+            return RedirectToAction("Index", "Admin");
         }
 
         // GET: Reservations/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Session["accessLevel"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Reservation reservation = db.Reservations.Find(id);
+                if (reservation == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(reservation);
             }
-            Reservation reservation = db.Reservations.Find(id);
-            if (reservation == null)
-            {
-                return HttpNotFound();
-            }
-            return View(reservation);
+
+            return RedirectToAction("Index", "Admin");
         }
 
         // GET: Reservations/Create
         public ActionResult Create()
         {
-            ViewBag.UserID = new SelectList(db.Users, "UserID", "Email");
-            return View();
+            if (Session["accessLevel"] != null)
+            {
+                ViewBag.UserID = new SelectList(db.Users, "UserID", "Email");
+                return View();
+            }
+
+            return RedirectToAction("Index", "Admin");
         }
 
         // POST: Reservations/Create
@@ -59,31 +79,41 @@ namespace ReserveIt.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ReservationID,UserID,RoomID,CheckIn,StayLength,NightlyRate,DiscountID")] Reservation reservation)
         {
-            if (ModelState.IsValid)
+            if (Session["accessLevel"] != null)
             {
-                db.Reservations.Add(reservation);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Reservations.Add(reservation);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+
+                ViewBag.UserID = new SelectList(db.Users, "UserID", "Email", reservation.UserID);
+                return View(reservation);
             }
 
-            ViewBag.UserID = new SelectList(db.Users, "UserID", "Email", reservation.UserID);
-            return View(reservation);
+            return RedirectToAction("Index", "Admin");
         }
 
         // GET: Reservations/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Session["accessLevel"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Reservation reservation = db.Reservations.Find(id);
+                if (reservation == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.UserID = new SelectList(db.Users, "UserID", "Email", reservation.UserID);
+                return View(reservation);
             }
-            Reservation reservation = db.Reservations.Find(id);
-            if (reservation == null)
-            {
-                return HttpNotFound();
-            }
-            ViewBag.UserID = new SelectList(db.Users, "UserID", "Email", reservation.UserID);
-            return View(reservation);
+
+            return RedirectToAction("Index", "Admin");
         }
 
         // POST: Reservations/Edit/5
@@ -93,29 +123,39 @@ namespace ReserveIt.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ReservationID,UserID,RoomID,CheckIn,StayLength,NightlyRate,DiscountID")] Reservation reservation)
         {
-            if (ModelState.IsValid)
+            if (Session["accessLevel"] != null)
             {
-                db.Entry(reservation).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(reservation).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.UserID = new SelectList(db.Users, "UserID", "Email", reservation.UserID);
+                return View(reservation);
             }
-            ViewBag.UserID = new SelectList(db.Users, "UserID", "Email", reservation.UserID);
-            return View(reservation);
+
+            return RedirectToAction("Index", "Admin");
         }
 
         // GET: Reservations/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Session["accessLevel"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Reservation reservation = db.Reservations.Find(id);
+                if (reservation == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(reservation);
             }
-            Reservation reservation = db.Reservations.Find(id);
-            if (reservation == null)
-            {
-                return HttpNotFound();
-            }
-            return View(reservation);
+
+            return RedirectToAction("Index", "Admin");
         }
 
         // POST: Reservations/Delete/5
@@ -123,10 +163,15 @@ namespace ReserveIt.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Reservation reservation = db.Reservations.Find(id);
-            db.Reservations.Remove(reservation);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["accessLevel"] != null)
+            {
+                Reservation reservation = db.Reservations.Find(id);
+                db.Reservations.Remove(reservation);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+
+            return RedirectToAction("Index", "Admin");
         }
 
         protected override void Dispose(bool disposing)
